I/o funcionando

+++ escalonadores.c
@@ -11,11 +11,69 @@
 #define MAX_PROG 10 
 #define INF 1000000
 
+int seg1 , seg2;
+int * prioridades_prontos;
+int index_prog_atual;
+int prog_io[MAX_PROG];
+int pid_atual;
+int index_prog_io, prioridade_prog_io;
+
+void intHandler(int dummy) {
+    printf("\nLimpando memoria compartilhada! \n");
+    shmdt (prioridades_prontos);
+    shmdt (&index_prog_atual);
+	shmctl (seg1, IPC_RMID, 0);
+
+	shmctl (seg2, IPC_RMID, 0);
+	exit(0);
+
+}
+
+void handler_io_r(int dummy)
+{
+	int n_pid = fork();
+	if(n_pid == 0)
+	{
+		printf("I/O recebido!\n");
+		prog_io[index_prog_atual] = 1;
+		index_prog_io = index_prog_atual;
+		kill(pid_atual, SIGSTOP);
+		sleep(4);
+		printf("I/O terminado!\n");
+		prog_io[index_prog_io] = 0;
+		exit(0);
+	}
+	else
+		return;
+}
+
+
+void handler_io_p(int dummy)
+{
+	int n_pid = fork();
+	if(n_pid == 0)
+	{
+		printf("I/O recebido!\n");
+		prioridade_prog_io = prioridades_prontos[index_prog_atual];
+		prioridades_prontos[index_prog_atual] = INF;
+		index_prog_io = index_prog_atual;
+		kill(pid_atual, SIGSTOP);
+		sleep(4);
+		printf("I/O terminado!\n");
+		prioridades_prontos[index_prog_io] = prioridade_prog_io;
+		exit(0);
+	}
+	else
+		return;
+}
+
 void round_robin(char * programas[MAX_PROG], int n)
 {
 	int i, pid[MAX_PROG], n_pid;
 	int finished[MAX_PROG];
 	int result, status;
+	signal(SIGINT,intHandler );
+	signal(SIGILL,handler_io_r);
 	for(i = 0; i < n; i++)
 		finished[i] = 0;
 
@@ -35,11 +93,15 @@ void round_robin(char * programas[MAX_PROG], int n)
 	}	
 
 	i = 0;
+
+
 	while(1)
 	{
-		if(finished[i] == 0)
+
+		if(finished[i] == 0 && prog_io[i] == 0)
 		{
 			kill(pid[i], SIGCONT);
+			pid_atual = pid[i];
 			sleep(5);
 			result = waitpid(pid[i], &status, WNOHANG);
 			if (result == 0) 
@@ -79,12 +141,11 @@ void prioridades(char * programas[MAX_PROG], int prioridades[MAX_PROG], int n)
 	int i, pid[MAX_PROG], n_pid, j, min;
 	int finished[MAX_PROG];
 
-	char * programas_prontos[MAX_PROG];
-	int seg1 , seg2;
-	int * prioridades_prontos;
-	int index_prog_atual = 0, index_prog_novo;
+	int index_prog_novo;
 	char aux[20];
 	int result, status;
+	signal(SIGINT,intHandler );
+	signal(SIGILL,handler_io_p );
 
 	seg1 = shmget(IPC_PRIVATE, sizeof(int)*MAX_PROG, IPC_CREAT | IPC_EXCL | S_IRUSR | S_IWUSR);
 	seg2 = shmget(IPC_PRIVATE, sizeof(int), IPC_CREAT | IPC_EXCL | S_IRUSR | S_IWUSR);
@@ -96,8 +157,6 @@ void prioridades(char * programas[MAX_PROG], int prioridades[MAX_PROG], int n)
 
 	for(i = 0 ; i < MAX_PROG; i++)
 	{
-		programas_prontos[i] = (char *) malloc(sizeof(char) * 10);
-		strcpy(programas_prontos[i], "");
 		prioridades_prontos[i] = INF;
 	}
 
@@ -119,10 +178,10 @@ void prioridades(char * programas[MAX_PROG], int prioridades[MAX_PROG], int n)
 
 	if(n_pid == 0)
 	{
+		index_prog_atual = 0;
 		for(i = 0 ; i < n; i++)
 		{
 			printf("Entrou na lista de pronto: %s com prioridade: %d\n", programas[i], prioridades[i]);
-			strcpy(programas_prontos[i], programas[i]);
 			prioridades_prontos[i] = prioridades[i];
 			index_prog_novo = i;
 			if(prioridades[index_prog_novo] < prioridades[index_prog_atual])
@@ -139,15 +198,13 @@ void prioridades(char * programas[MAX_PROG], int prioridades[MAX_PROG], int n)
 		{
 
 			i = menor_prioridade(prioridades_prontos, n);
-			//printf("%d %d %d\n", prioridades_prontos[0], prioridades_prontos[1], prioridades_prontos[2]);
-			//scanf("%d", &min);
 			if(i == INF)
 				return;
 
 			if(finished[i] == 0)
 			{
 				kill(pid[i], SIGCONT);
-				//printf("PID1: %d\n", pid[i]);
+				pid_atual = pid[i];
 				index_prog_atual = i;
 				result = waitpid(pid[i], &status, WNOHANG);
